Add files via upload
My urlscrape is just the old one with some experimentation. Since we decided to use a different method, I figured I would focus on the algorithm for now. 
The algorithm is really quite simple. Just sort based on a calculated priority. I have only tested it on a single term, so I will have to add additional functionality to take years, time periods, and scheduling conflicts (when two classes with same priority are only offered at the same time) into account. However, I think that should be pretty easy to deal with as well.