Removed Archive Folder
Removed unnecessary Archive folder.
Got rid of the Section data structure as it was unneccesarry.
Scraper now uses API instead of html.
Still need to finish functions for prerequistes, terms offered, and
priority calculation.
Will add a variabel in the Class object to store if class is "night
friendly." Also a function will be added to get and calculate said
variable.