Scraper Update and Cataclysm
-Degrees now store Concentrations by index instead of storing the
Concentration objects themselves. This is for database compatability.
-Concentrations now store their Classes in the same fassion as above.
-Added 'OR' variable to Class Objects to handle when there is an
either/or Class path option.
-Wrote most of the regular expressions to succesfully parse the
PREREQUISTES for applicable classes.
-While finishing 'get_prereqs' an update on the DePaul website was
pushed through. This resulted in some of the urls being changed, which I
fixed, but it also seems the data is no longer available in a json
format and the scraper now gives a 'No JSON object could be decoded'
error.